# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GIAM (General Identity and Access Manager)",
    "version": "1.0.0",
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "clientBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "mtls": {
        "type": "mutualTLS"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "description": "RFC 9457 problem details.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI identifying the problem type.",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary.",
            "example": "Not Found"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code.",
            "example": 404
          },
          "detail": {
            "type": "string",
            "description": "An explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "description": "A URI identifying this occurrence."
          }
        },
        "title": "Problem",
        "example": {
          "type": "about:blank",
          "title": "Not Found",
          "status": 404
        }
      },
      "def-1": {
        "type": "object",
        "description": "RFC 6749 section 5.2 error response.",
        "required": [
          "error"
        ],
        "additionalProperties": true,
        "properties": {
          "error": {
            "type": "string",
            "description": "The RFC 6749 error code.",
            "example": "invalid_request"
          },
          "error_description": {
            "type": "string",
            "description": "Human-readable detail."
          },
          "error_uri": {
            "type": "string",
            "description": "A page describing the error."
          }
        },
        "title": "OAuthError",
        "example": {
          "error": "invalid_request",
          "error_description": "grant_type is required"
        }
      }
    }
  },
  "paths": {
...
    "/realms/{realm}/protocol/openid-connect/auth": {
      "post": {
        "operationId": "authorize",
        "summary": "Authorization endpoint",
        "tags": [
          "oauth"
        ],
        "description": "Standard-defined: RFC 6749 section 4.1 and RFC 7636 (PKCE). Exchanges an established session for a single-use authorization code bound to the client, the redirect URI and the PKCE challenge. It never accepts a credential: a request with no session is told one is required rather than being prompted here.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "client_id",
                  "redirect_uri",
                  "response_type",
                  "session_id"
                ],
                "additionalProperties": false,
                "properties": {
                  "client_id": {
                    "type": "string",
                    "example": "orders-web"
                  },
                  "redirect_uri": {
                    "type": "string",
                    "example": "https://app.example/callback"
                  },
                  "response_type": {
                    "type": "string",
                    "enum": [
                      "code"
                    ]
                  },
                  "scope": {
                    "type": "string",
                    "example": "openid profile"
                  },
                  "state": {
                    "type": "string"
                  },
                  "nonce": {
                    "type": "string"
                  },
                  "code_challenge": {
                    "type": "string"
                  },
                  "code_challenge_method": {
                    "type": "string",
                    "enum": [
                      "S256"
                    ]
                  },
                  "session_id": {
                    "type": "string",
                    "description": "The session established at sign-in."
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "example": "acme",
            "in": "path",
            "name": "realm",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "The authorization code, and the state to echo back.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The authorization code, and the state to echo back.",
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "redirect_uri"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "redirect_uri": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "a1b2…",
                  "state": "xyz",
                  "redirect_uri": "https://app.example/callback"
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid or the scope is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          },
          "401": {
            "description": "No live session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
...
    "/realms/{realm}/federation/{provider}/start": {
      "get": {
        "operationId": "startFederation",
        "summary": "Begin signing in through an upstream provider",
        "tags": [
          "authentication"
        ],
        "description": "Standard-defined: OpenID Connect Core 1.0 authentication request, made by this authority as a relying party of the upstream provider. Public, because it is the first step of a sign-in and the person has nothing to present yet.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "realm",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "example": "entra",
            "in": "path",
            "name": "provider",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Where to send the browser, and the state to expect back.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Where to send the browser, and the state to expect back.",
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "authorizationUrl",
                    "state"
                  ],
                  "properties": {
                    "authorizationUrl": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "authorizationUrl": "https://login.example/authorize?…",
                  "state": "a1b2c3"
                }
              }
            }
          },
          "404": {
            "description": "No such realm or provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "503": {
            "description": "The provider is configured but not reachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/realms/{realm}/federation/{provider}/callback": {
      "post": {
        "operationId": "completeFederation",
        "summary": "Complete a sign-in through an upstream provider",
        "tags": [
          "authentication"
        ],
        "description": "Standard-defined: OpenID Connect Core 1.0 code exchange, with the upstream id token verified against the provider's published keys before any claim in it is believed. Public for the same reason the sign-in route is: the code IS the credential being presented.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "state"
                ],
                "additionalProperties": false,
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string",
                    "description": "The value this authority issued when the flow started."
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "realm",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "provider",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "The local principal, and the session established for it.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The local principal, and the session established for it.",
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "subjectId",
                    "sessionId"
                  ],
                  "properties": {
                    "subjectId": {
                      "type": "string"
                    },
                    "sessionId": {
                      "type": "string"
                    },
                    "userName": {
                      "type": "string"
                    },
                    "provisioned": {
                      "type": "boolean",
                      "description": "True when this sign-in created the principal."
                    }
                  }
                },
                "example": {
                  "subjectId": "sub-9f21",
                  "sessionId": "ses-4c1f",
                  "userName": "ada",
                  "provisioned": false
                }
              }
            }
          },
          "401": {
            "description": "The upstream response did not verify.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "No such realm or provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8085",
      "description": "private, service to service"
    }
  ],
...
}